Give frequently accessed entries more chance to stay alive.
When adding a new element to the cuckoo filter, if there is not enough
space in the filter, a random element gets removed. But it is not desired
for frequently accessed items to get randomly removed from the cache. we
want to keep frequently accessed items in the cache as long as possible.
To fix this, on each cache hit we add the accessed element to the filter,
even if it already exists. Using this technique, for frequently accessed
items we will have duplicate entries in the filter, hence giving that
item more chance to stay in the cache when items are being randomly
removed.